+++ R/perf_mod.R
@@ -1,78 +1,84 @@
 #' Bayesian Analysis of Resampling Statistics
 #'
-#' Bayesian analysis used here to answer the question: "when
-#'  looking at resampling results, are the differences between
-#'  models 'real?'" To answer this, a model can be created were the
-#'  _outcome_ is the resampling statistics (e.g. accuracy or RMSE).
-#'  These values are explained by the model types. In doing this, we
-#'  can get parameter estimates for each model's affect on
-#'  performance and make statistical (and practical) comparisons
-#'  between models.
+#'   Bayesian analysis used here to answer the question: "when looking at
+#'  resampling results, are the differences between models 'real?'" To answer
+#'  this, a model can be created were the _outcome_ is the resampling statistics
+#'  (e.g. accuracy or RMSE). These values are explained by the model types. In
+#'  doing this, we can get parameter estimates for each model's affect on
+#'  performance and make statistical (and practical) comparisons between models.
 #'
 #' @param object A data frame or an `rset` object (such as
-#'  [rsample::vfold_cv()]) containing the `id` column(s) and at least
-#'  two numeric columns of model performance statistics (e.g.
-#'  accuracy). Additionally, an object from `caret::resamples`
-#'  can be used.
-#' @param ... Additonal arguments to pass to [rstanarm::stan_glmer()]
-#'  such as `verbose`, `prior`, `seed`, `family`, etc.
+#'  [rsample::vfold_cv()]) containing the `id` column(s) and at least two
+#'  numeric columns of model performance statistics (e.g. accuracy).
+#'  Additionally, an object from `caret::resamples` can be used.
+#' @param formula An optional model formula to use for the Bayesian hierarchical model
+#' (see Details below).
+#' @param ... Additional arguments to pass to [rstanarm::stan_glmer()] such as
+#'  `verbose`, `prior`, `seed`, `family`, etc.
 #' @return An object of class `perf_mod`.
-#' @details These functions can be used to process and analyze
-#'  matched resampling statistics from different models using a
-#'  Bayesian generalized linear model with effects for the model and
-#'  the resamples.
+#' @details These functions can be used to process and analyze matched
+#'  resampling statistics from different models using a Bayesian generalized
+#'  linear model with effects for the model and the resamples.
 #'
-#' By default, a generalized linear model with Gaussian error and
-#'  an identity link is fit to the data and has terms for the
-#'  predictive model grouping variable. In this way, the performance
-#'  metrics can be compared between models.
+#'   By default, a generalized linear model with Gaussian error and an identity
+#'  link is fit to the data and has terms for the predictive model grouping
+#'  variable. In this way, the performance metrics can be compared between
+#'  models.
 #'
-#' Additionally, random effect terms are also used. For most
-#'  resampling methods (except repeated _V_-fold cross-validation),
-#'  a simple random intercept model its used with an exchangeable
-#'  (i.e. compound-symmetric) variance structure. In the case of
-#'  repeated cross-validation, two random intercept terms are used;
-#'  one for the repeat and another for the fold within repeat. These
-#'  also have exchangeable correlation structures.
+#'   Additionally, random effect terms are also used. For most resampling
+#'  methods (except repeated _V_-fold cross-validation), a simple random
+#'  intercept model its used with an exchangeable (i.e. compound-symmetric)
+#'  variance structure. In the case of repeated cross-validation, two random
+#'  intercept terms are used; one for the repeat and another for the fold within
+#'  repeat. These also have exchangeable correlation structures.
 #'
-#' The above model specification assumes that the variance in the
-#'  performance metrics is the same across models. However, this is
-#'  unlikely to be true in some cases. For example, for simple
-#'  binomial accuracy, it well know that the variance is highest
-#'  when the accuracy is near 50 percent. When the argument
-#'  `hetero_var = TRUE`, the variance structure uses random
-#'  intercepts for each model term. This may produce more realistic
-#'  posterior distributions but may take more time to converge.
+#'   The above model specification assumes that the variance in the performance
+#'  metrics is the same across models. However, this is unlikely to be true in
+#'  some cases. For example, for simple binomial accuracy, it well know that the
+#'  variance is highest when the accuracy is near 50 percent. When the argument
+#'  `hetero_var = TRUE`, the variance structure uses random intercepts for each
+#'  model term. This may produce more realistic posterior distributions but may
+#'  take more time to converge.
 #'
-#' Also, as shown in the package vignettes, the Gaussian assumption
-#'  make be unrealistic. In this case, there are at least two
-#'  approaches that can be used. First, the outcome statistics can
-#'  be transformed prior to fitting the model. For example, for
-#'  accuracy, the logit transformation can be used to convert the
-#'  outcome values to be on the real line and a model is fit to
-#'  these data. Once the posterior distributions are computed, the
-#'  inverse transformation can be used to put them back into the
-#'  original units. The `transform` argument can be used to do this.
+#'  Examples of the default formulas are:
 #'
-#' The second approach would be to use a different error
-#'  distribution from the exponential family. For RMSE values, the
-#'  Gamma distribution may produce better results at the expense of
-#'  model computational complexity. This can be achieved by passing
-#'  the `family` argument to `perf_mod` as one might with the
-#'  `glm` function.
-#' @examples
-#' # Example objects from the "Getting Started" vignette at
-#' #  https://topepo.github.io/tidyposterior/articles/Getting_Started.html
+#'  \preformatted{
+#'  # One ID field and common variance:
+#'    statistic ~ model + (model | id)
 #'
-#' # File for pre-run model is at
-#' ex_dat <- "https://bit.ly/2OJdvl1"
+#'  # One ID field and heterogeneous variance:
+#'    statistic ~ model + (model + 0 | id)
 #'
-#' # load(load(url(ex_dat))
+#'  # Repeated CV (id = repeat, id2 = fold within repeat)
+#'  # with a common variance:
+#'   statistic ~ model + (model | id2/id)
 #'
-#' # roc_model
+#'  # Repeated CV (id = repeat, id2 = fold within repeat)
+#'  # with a heterogeneous variance:
+#'   statistic ~ model + (model + 0| id2/id)
 #'
-#' # Summary method shows the underlying `stan` model
-#' # summary(roc_model)
+#'  # Default for unknown resampling method and
+#'  # multiple ID fields:
+#'   statistic ~ model + (model | idN/../id)
+#'  }
+#'
+#'  Custom formulas should use `statistic` as the outcome variable and `model`
+#'  as the factor variable with the model names.
+#'
+#'   Also, as shown in the package vignettes, the Gaussian assumption make be
+#'  unrealistic. In this case, there are at least two approaches that can be
+#'  used. First, the outcome statistics can be transformed prior to fitting the
+#'  model. For example, for accuracy, the logit transformation can be used to
+#'  convert the outcome values to be on the real line and a model is fit to
+#'  these data. Once the posterior distributions are computed, the inverse
+#'  transformation can be used to put them back into the original units. The
+#'  `transform` argument can be used to do this.
+#'
+#'   The second approach would be to use a different error distribution from the
+#'  exponential family. For RMSE values, the Gamma distribution may produce
+#'  better results at the expense of model computational complexity. This can be
+#'  achieved by passing the `family` argument to `perf_mod` as one might with
+#'  the `glm` function.
 #' @export
 perf_mod <- function(object, ...)
   UseMethod("perf_mod")
@@ -91,21 +97,15 @@ perf_mod.default <- function(object, ...)
 
 #' @rdname perf_mod
 #' @param transform An named list of transformation and inverse
-#'  transformation fuctions. See [logit_trans()] as an example.
+#'  transformation functions. See [logit_trans()] as an example.
 #' @param hetero_var A logical; if `TRUE`, then different
 #'  variances are estimated for each model group. Otherwise, the
 #'  same variance is used for each group. Estimating heterogeneous
 #'  variances may slow or prevent convergence.
 #' @export
-#' @importFrom dplyr filter select mutate %>%
-#' @importFrom rsample gather.rset pretty.group_vfold_cv
-#' @importFrom rsample pretty.bootstraps pretty.nested_cv
-#' @importFrom rsample pretty.mc_cv pretty.rolling_origin
-#' @importFrom rsample pretty.loo_cv pretty.vfold_cv
-#' @importFrom rstanarm stan_glmer
-#' @importFrom rlang !!
+
 perf_mod.rset <-
-  function(object, transform = no_trans, hetero_var = FALSE, ...) {
+  function(object, transform = no_trans, hetero_var = FALSE, formula = NULL, ...) {
     check_trans(transform)
     rset_type <- try(pretty(object), silent = TRUE)
     if(inherits(rset_type, "try-error"))
@@ -123,16 +123,12 @@ perf_mod.rset <-
 
     ## Make a formula based on resampling type (repeatedcv, rof),
     ## This could be done with more specific classes
+    id_cols <- grep("(^id)|(^id[1-9]$)", names(object), value = TRUE)
+    formula <- make_formula(id_cols, hetero_var, formula)
 
     model_names <- unique(as.character(resamples$model))
 
-    if (hetero_var) {
-      mod <- stan_glmer(statistic ~  model + (model + 0 | id),
-                        data = resamples, ...)
-    } else {
-      mod <- stan_glmer(statistic ~  model + (1 | id),
-                        data = resamples, ...)
-    }
+    mod <- stan_glmer(formula, data = resamples, ...)
 
     res <- list(stan = mod,
                 hetero_var = hetero_var,
@@ -144,9 +140,42 @@ perf_mod.rset <-
     res
   }
 
+make_formula <- function (ids, hetero_var, formula) {
+  if (is.null(formula)) {
+    ids <- sort(ids)
+    p <- length(ids)
+    if(p > 1) {
+      msg <-
+        paste0("There were multiple resample ID columns in the data. It is ",
+               "unclear what the model formula should be for the hierarchical ",
+               "model. This analysis used the formula: ")
+      nested <- paste0(rev(ids), collapse = "/")
+      if (hetero_var) {
+        f_chr <- paste0("statistic ~  model + (model + 0 |", nested, ")")
+        f <- as.formula(f_chr)
+      } else {
+        f_chr <- paste0("statistic ~  model + (1 |", nested, ")")
+        f <- as.formula(f_chr)
+      }
+      msg <- paste0(msg, rlang::expr_label(f),
+                    " The `formula` arg can be used to change this value.")
+      rlang::warn(msg)
+    } else {
+      if (hetero_var) {
+        f <- statistic ~  model + (model + 0 | id)
+      } else {
+        f <- statistic ~  model + (1 | id)
+      }
+    }
+  } else {
+    f <- formula
+  }
+  attr(f, ".Environment") <- rlang::base_env()
+  f
+}
+
+
 #' @export
-#' @exportMethod perf_mod vfold_cv
-#' @importFrom rsample vfold_cv
 #' @rdname perf_mod
 perf_mod.vfold_cv <-
   function(object, transform = no_trans, hetero_var = FALSE, ...) {
@@ -191,11 +220,6 @@ perf_mod.vfold_cv <-
     res
   }
 
-
-
-#' @importFrom utils globalVariables
-utils::globalVariables(c("id", "model", "splits", "statistic", "Resample"))
-
 #' @export
 print.perf_mod <- function(x, ...) {
   cat("Bayesian Analysis of Resampling Results\n")
@@ -214,10 +238,8 @@ summary.perf_mod <- function(object, ...) {
 
 
 #' @export
-#' @importFrom stats setNames
-#' @importFrom purrr map_chr
 #' @rdname perf_mod
-#' @param metric A single character value for the statstic from
+#' @param metric A single character value for the statistic from
 #'  the `resamples` object that should be analyzed.
 perf_mod.resamples <-
   function(object,
@@ -261,27 +283,14 @@ perf_mod.data.frame <-
   function(object,
            transform = no_trans,
            hetero_var = FALSE,
+           formula = NULL,
            ...) {
-    id_cols <- grep("(^id)|(^id[1-9]$)",
-                    names(object),
-                    value = TRUE)
+    id_cols <- grep("(^id)|(^id[1-9]$)", names(object), value = TRUE)
     if (length(id_cols) == 0)
       stop("One or more `id` columns are required.", call. = FALSE)
-
-    if (length(id_cols) > 1) {
-      warning(
-        "Since no specific resampling method is known,",
-        "the ID variables are collapsed into one column.",
-        call. = FALSE
-      )
-      tmp <- apply(object[, id_cols], 1, paste0, collapse = "-")
-      for (i in id_cols)
-        object[, i] <- NULL
-      object$id <- tmp
-    }
 
     class(object) <- c("rset", class(object))
 
     perf_mod(object, transform = transform,
-             hetero_var = hetero_var, ...)
+             hetero_var = hetero_var, formula = formula, ...)
   }